Cleanup labels_of_service and label_sources_of_service API
Both functions required the service labels as input parameters
since a short time. This was only done this way to be able to
re-use the service labels read from the autochecks of the host
and a bad API for most use cases.

The source label loading logic should be transparent to the call
sites, just like the caching of the persistance logic.

This commit restores the previous API where the call sites only
need the host name and service description. All other needed
information will be gathered by the LabelManager on it's own.

Currently the LabelManager._discovered_labels_of_service() method
is hacked to just make it work. The next commit will move the
autochecks loading logic to cmk.utils, to make it available to the
all components of Checkmk and wrap the code into a manager class
that cares about caching the already read data during runtime.

Change-Id: I75bf9d4296816fca86fea8c359d33ec48bcff574