Cleanup special case for binary rulesets
Previous to 1.6 the binary rulesets had no value and could not have a
valuespec for the input. In the new rule format the binary rulesets also
have a value (True/False). This makes it possible to give these rulesets
valuespecs that can represent these values and to reduce some special
cases.

CMK-2427

Change-Id: Ib897323d4c24982a3bdb4d08295305446cb82f64